Refactor IKEA five button remote quirks (#2155)
* Merge fivebtnremote.py and fivebtnremotezha.py

* Rename quirk classes

* Share device automation triggers whilst keeping originals

* Share device automation triggers, remove `"param1"=537` for IkeaTradfriRemote2

* Update signature comment for IkeaTradfriRemote3

* Update doc comments
